+++ next-tavla/app/(admin)/edit/[id]/components/Buttons/Copy.tsx
@@ -1,11 +1,9 @@
 'use client'
 import { useToast } from '@entur/alert'
 import { Button, IconButton } from '@entur/button'
-import { CopyIcon, ExternalIcon } from '@entur/icons'
+import { CopyIcon } from '@entur/icons'
 import { Tooltip } from '@entur/tooltip'
 import { useLink } from 'hooks/useLink'
-import Link from 'next/link'
-import React from 'react'
 
 function Copy({ type, bid }: { type?: 'button' | 'icon'; bid?: string }) {
     const { addToast } = useToast()
@@ -35,36 +33,4 @@ function Copy({ type, bid }: { type?: 'button' | 'icon'; bid?: string }) {
         </Tooltip>
     )
 }
-
-function Open({ type, bid }: { type?: 'button' | 'icon'; bid?: string }) {
-    const link = useLink(bid)
-    if (type === 'button') {
-        return (
-            <Button
-                variant="primary"
-                as={Link}
-                aria-label="Åpne tavle"
-                href={link ?? '/'}
-                target="_blank"
-            >
-                Åpne Tavle
-                <ExternalIcon />
-            </Button>
-        )
-    }
-
-    return (
-        <Tooltip content="Åpne tavle" placement="bottom">
-            <IconButton
-                as={Link}
-                aria-label="Åpne tavle"
-                href={link ?? '/'}
-                target="_blank"
-            >
-                <ExternalIcon />
-            </IconButton>
-        </Tooltip>
-    )
-}
-
-export { Open, Copy }
+export { Copy }
